Block DD from unauthorized users
update unit tests
